requirements.txt
# Core AI & Data Dependencies
PyPDF2==3.0.1
python-docx==1.1.0
scikit-learn==1.3.2
numpy==1.26.0 
python-dotenv==1.0.0

# Testing
pytest==7.4.3

# Optional: For testing API
fastapi==0.104.1
uvicorn==0.24.0